Phidata_Agents/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
openai
python-multipart
fastapi
uvicorn


